Move feed page slice to common
