+++ src/pages/feed-page/components/mobile/FeedPageContent.tsx
@@ -5,6 +5,7 @@ import cn from 'classnames'
 import { Name } from 'common/models/Analytics'
 import FeedFilter from 'common/models/FeedFilter'
 import Status from 'common/models/Status'
+import { feedActions } from 'common/store/pages/feed/lineup/actions'
 import Header from 'components/header/mobile/Header'
 import { HeaderContext } from 'components/header/mobile/HeaderContextProvider'
 import Lineup from 'components/lineup/Lineup'
@@ -12,7 +13,6 @@ import MobilePageContainer from 'components/mobile-page-container/MobilePageCont
 import { useMainPageHeader } from 'components/nav/store/context'
 import PullToRefresh from 'components/pull-to-refresh/PullToRefresh'
 import useAsyncPoll from 'hooks/useAsyncPoll'
-import { feedActions } from 'pages/feed-page/store/lineups/feed/actions'
 import { FeedPageContentProps } from 'pages/feed-page/types'
 import { make, useRecord } from 'store/analytics/actions'
 import { BASE_URL, FEED_PAGE } from 'utils/route'
