+++ src/pages/feed-page/store/sagas.ts
@@ -3,9 +3,9 @@ import { call, put, take, fork, takeEvery } from 'redux-saga/effects'
 import { FollowSource } from 'common/models/Analytics'
 import { ID } from 'common/models/Identifiers'
 import { fetchUsers } from 'common/store/cache/users/sagas'
+import * as discoverActions from 'common/store/pages/feed/actions'
+import { feedActions } from 'common/store/pages/feed/lineup/actions'
 import * as socialActions from 'common/store/social/users/actions'
-import * as discoverActions from 'pages/feed-page/store/actions'
-import { feedActions } from 'pages/feed-page/store/lineups/feed/actions'
 import { fetchSuggestedFollowUserIds } from 'pages/sign-on/store/sagas'
 import { waitForBackendSetup } from 'store/backend/sagas'
