+++ src/store/lineup/lineupForRoute.js
@@ -1,8 +1,8 @@
 import { matchPath } from 'react-router'
 
+import { getDiscoverFeedLineup } from 'common/store/pages/feed/selectors'
 import { getLineup } from 'common/store/pages/track/selectors'
 import { getCollectionTracksLineup } from 'pages/collection-page/store/selectors'
-import { getDiscoverFeedLineup } from 'pages/feed-page/store/selectors'
 import { getHistoryTracksLineup } from 'pages/history-page/store/selectors'
 import { getProfileTracksLineup } from 'pages/profile-page/store/selectors'
 import { getSavedTracksLineup } from 'pages/saved-page/store/selectors'
