+++ src/store/reducers.js
@@ -2,6 +2,7 @@ import { connectRouter } from 'connected-react-router'
 import { combineReducers } from 'redux'
 
 import { reducers as clientStoreReducers } from 'common/store'
+import feed from 'common/store/pages/feed/reducer'
 import remoteConfig from 'common/store/remote-config/slice'
 import artistRecommendations from 'components/artist-recommendations/store/slice'
 import changePassword from 'components/change-password/store/slice'
@@ -22,7 +23,6 @@ import deleted from 'pages/deleted-page/store/slice'
 import exploreCollections from 'pages/explore-page/store/collections/slice'
 import explore from 'pages/explore-page/store/reducer'
 import favorites from 'pages/favorites-page/store/reducers'
-import feed from 'pages/feed-page/store/reducer'
 import followers from 'pages/followers-page/store/reducers'
 import following from 'pages/following-page/store/reducers'
 import history from 'pages/history-page/store/reducer'
